Senors work. Actuators?

+++ trunk/src/com/traclabs/biosim/server/sensor/framework/GenericSensorImpl.java
@@ -20,7 +20,7 @@ public abstract class GenericSensorImpl extends BioModuleImpl implements
 
     protected void notifyListeners(){
     	if (isBionetEnabled())
-    		myBionetNode.setResourceValue(BionetUtils.RESOURCE_ID, Float.valueOf(getValue()), new Timeval());
+    		myBionetNode.setResourceValue(BionetUtils.RESOURCE_ID, Double.valueOf(getValue()), new Timeval());
     }
 
     public float getValue() {
